pyproject.toml
[tool.poetry]
name = "ms-server"
version = "0.1.0"
description = ""
authors = ["SergeevDS"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.5"
django = "^5.1.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
